+++ tasks/task_3/kalinin_a_gradient/CMakeLists.txt
