+++ tasks/task_3/kalinin_a_gradient/main.cpp
@@ -1,0 +1,23 @@
+// Copyright 2023 Kalinin Alexandr
+#include <gtest/gtest.h>
+#include "./gradient.h"
+}
+
+int main(int argc, char** argv) {
+    int result_code = 0;
+    ::testing::InitGoogleTest(&argc, argv);
+    ::testing::TestEventListeners& listeners =
+        ::testing::UnitTest::GetInstance()->listeners();
+
+    if (MPI_Init(&argc, &argv) != MPI_SUCCESS)
+        MPI_Abort(MPI_COMM_WORLD, -1);
+    int rank;
+    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+    if (rank != 0) {
+        delete listeners.Release(listeners.default_result_printer());
+    }
+    result_code = RUN_ALL_TESTS();
+    MPI_Finalize();
+
+    return result_code;
+}
